+++ project_market/src/main/java/pjmarket/controller/AddrController.java
@@ -1,0 +1,5 @@
+package pjmarket.controller;
+
+public class AddrController {
+
+}
