+++ project_market/src/main/java/pjmarket/controller/ProductController.java
@@ -188,5 +188,4 @@ public class ProductController {
 		return "main/product_detail";
 	}
 
-	
 }
