+++ project_market/src/main/webapp/WEB-INF/views/fix/header.jsp
@@ -887,10 +887,10 @@ $(document).ready(function() {
 					</div>
 				</li>
 				<li><div class="menu-box active">
-						<a href="">Rangking</a>
+						<a href="off_list.do">공지사항</a>
 					</div></li>
 				<li><div class="menu-box active">
-						<a href="qna_boardlist.do?page=1">Q&A</a>
+						<a href="#">##</a>
 					</div></li>
 				<li><div class="menu-box active">
 						<a href="review_writeform">Review</a>
