+++ project_market/src/main/webapp/WEB-INF/views/main/mainpage.jsp
@@ -4,6 +4,7 @@
 <%@ include file="../fix/header.jsp"%>
 <section class="main">
 	<div class="main-img">
+		활성화된 세션: ${member_id }
 		<div class=welcome>
 <!-- 		<div class="bubbles">
 			<h1 id="welcome_wording">WELCOME<br>GAGUMARKET</h1>
